Homework 4 done(moved Bet to another package)
