+++ src/hw4/Bet/Bet.java
@@ -1,4 +1,4 @@
-package hw4;
+package hw4.Bet;
 
 import java.util.Objects;
