+++ src/hw4/Bet/BetException.java
@@ -1,4 +1,4 @@
-package hw4;
+package hw4.Bet;
 
 public class BetException extends Exception {
